+++ tools/update-post.js
@@ -3,13 +3,28 @@ import { htmlToMarkdown } from "~/utils/html-to-markdown.js";
 import { stringify } from "@std/toml";
 import { extractToml as extract } from "@std/front-matter";
 
-import { kv } from "./utils/production-kv.js";
 import { markdownToHTML } from "../utils/markdown-to-html.js";
+import { invariant } from "../utils/invariant.js";
 
-const posts = await Array.fromAsync(
-	kv.list({ prefix: ["posts"] }),
-	(entry) => entry.value,
-);
+const MIGRATION_PATH = Deno.env.get("MIGRATION_PATH");
+invariant(MIGRATION_PATH);
+
+const MIGRATION_TOKEN = Deno.env.get("MIGRATION_TOKEN");
+invariant(MIGRATION_TOKEN);
+
+const PRODUCTION_URL = Deno.env.get("PRODUCTION_URL");
+invariant(PRODUCTION_URL);
+
+const ENDPOINT = new URL(`${MIGRATION_PATH}/posts`, PRODUCTION_URL);
+
+let response = await fetch(ENDPOINT, {
+	method: "GET",
+	headers: {
+		authorization: `Bearer ${MIGRATION_TOKEN}`,
+	},
+});
+
+const posts = await response.json();
 const slugs = posts.map((post) => post.slug).join("\n");
 
 const echo = new Deno.Command("echo", { args: [slugs] });
@@ -32,7 +47,18 @@ if (!fzfOutput.success) {
 }
 
 const slug = new TextDecoder().decode(fzfOutput.stdout).trim();
-const post = posts.find((p) => p.slug === slug);
+
+response = await fetch(`${ENDPOINT}/${slug}?by=slug`, {
+	method: "GET",
+	headers: {
+		authorization: `Bearer ${MIGRATION_TOKEN}`,
+	},
+});
+
+if (!response.ok) throw `Post not found with slug: ${slug}`;
+
+const post = await response.json();
+
 const { html, ...postWithoutHtml } = post;
 
 const markdown = String(await htmlToMarkdown(html));
@@ -62,15 +88,23 @@ const updatedPost = {
 	updatedAt: new Date().toISOString(),
 };
 
-const url = `https://knowler.dev/blog/${updatedPost.slug}`;
+const url = `${PRODUCTION_URL}/blog/${updatedPost.slug}`;
 
-await kv.set(["posts", post.id], updatedPost);
+response = await fetch(`${ENDPOINT}/${post.id}/update`, {
+	method: "POST",
+	body: JSON.stringify(updatedPost),
+	headers: {
+		"content-type": "application/json",
+		authorization: `Bearer ${MIGRATION_TOKEN}`,
+	},
+});
 
-// Bust the content cache
-await import("./utils/bust-content-cache.js");
+if (response.ok) {
+	await Deno.remove(fileName);
+	console.log(
+		`Successfully updated post: ${url}`,
+	);
+} else {
+	console.error(`Could not update post. Edits are here: ${fileName}`)
+}
 
-await Deno.remove(fileName);
-
-console.log(
-	`Successfully updated post: ${url}`,
-);
